+++ src/extensions/external_links.rs
@@ -181,7 +181,7 @@ impl ExternalLinks {
 
     fn visit_figure_target(&self, f: &mut crate::ast::Figure) {
         use crate::ast::FigureTarget;
-        match &mut f.target {
+        match &mut *f.target {
             FigureTarget::BlockQuote(b) => {
                 for child in &mut b.children {
                     self.visit_block(child);